Create a utility document instead of having it all be in one placce
